Refactor out the script into the reactive module
This branch strips most of the flakey and often misunderstood bootstrap.sh file. in favor of the reactive module code and python native tooling.

The only leftover from bootstrap.sh is the etcdctl configuration run command, due to subprocess stripping the quotes and malforming the json.

This PR depends on a new cut of charms.docker: https://github.com/juju-solutions/charms.docker/pull/23