Fix for etcd-retry scenario
This catches the error emitted when the etcd unit is still configuring
and not available. The self registration code would return > 0 and cause
the charm to halt execution causing juju-wait and conjure-up to fail
deployments.

This shamelessly takes the retry decorator that was piloted in the
kubernetes charms and embeds it in lib/charms/flannel/common.py

Closes https://github.com/juju-solutions/bundle-canonical-kubernetes/issues/312