Adjust the etcd state to wait until we've gotten certificates.
This will ensure that the etcd daemon has all the pre-req relationships
and it *should* be ready to receive our request. If not the retry catch
will delay up to a total of 1 minute while it attempts self
registration. otherwise it returns execution and should invoke again on
the next run (verified)