Fix for etcd-retry scenario (#32)
* Fix for etcd-retry scenario

This catches the error emitted when the etcd unit is still configuring
and not available. The self registration code would return > 0 and cause
the charm to halt execution causing juju-wait and conjure-up to fail
deployments.

This shamelessly takes the retry decorator that was piloted in the
kubernetes charms and embeds it in lib/charms/flannel/common.py

Closes https://github.com/juju-solutions/bundle-canonical-kubernetes/issues/312

* Adjust the etcd state to wait until we've gotten certificates.

This will ensure that the etcd daemon has all the pre-req relationships
and it *should* be ready to receive our request. If not the retry catch
will delay up to a total of 1 minute while it attempts self
registration. otherwise it returns execution and should invoke again on
the next run (verified)