Add subnet-len configuration option.
By default flannel chops the configured cidr in /24 subnets that are
assigned to each node, this allows kubelet to run up to 254 hosts, for
clusters where the machines are large enough to host more than 254 pods
or have a large number of create-destroy pod operations (e.g. a CI/CD
pipeline) would produce an starvation of available IP addresses.

This configuration option will allow flannel to be configured with
a different size for each subnet. When this configuration option
changes the cni0 bridge is destroyed and flannel restarted.

Fixes [lp:1930087](https://bugs.launchpad.net/charm-flannel/+bug/1930087)